requirements.txt
fastapi
uvicorn[standard]
opencv-python
Pillow
numpy
tqdm
sentence-transformers